requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
idna==2.10
image==1.5.32
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
pyfakewebcam==0.1.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
